Update CLI to match API

+++ harness/determined/cli/user_groups.py
@@ -3,13 +3,11 @@ from argparse import Namespace
 from collections import namedtuple
 from typing import Any, Dict, List, Optional
 
-from determined.cli import setup_session, default_pagination_args, require_feature_flag
+from determined.cli import default_pagination_args, render, require_feature_flag, setup_session
 from determined.common import api
 from determined.common.api import authentication, bindings
 from determined.common.declarative_argparse import Arg, Cmd
 from determined.common.experimental import session
-
-from . import render
 
 v1UserHeaders = namedtuple(
     "v1UserHeaders",
@@ -18,11 +16,13 @@ v1UserHeaders = namedtuple(
 
 v1GroupHeaders = namedtuple(
     "v1GroupHeaders",
-    ["groupId", "name"],
+    ["groupId", "name", "numMembers"],  # numMembers
 )
 
-rbac_flag_disabled_message = ("User groups commands require the Determined Enterprise Edition " +
-                              "and the Master Configuration option security.rbac.enabled.")
+rbac_flag_disabled_message = (
+    "User groups commands require the Determined Enterprise Edition "
+    + "and the Master Configuration option security.rbac.enabled."
+)
 
 
 @authentication.required
@@ -40,21 +40,33 @@ def create_group(args: Namespace) -> None:
 
 
 @authentication.required
+@require_feature_flag("rbacEnabled", rbac_flag_disabled_message)
 def list_groups(args: Namespace) -> None:
-    body = bindings.v1GetGroupsRequest(offset=args.offset, limit=args.limit)
-    resp = bindings.post_GetGroups(setup_session(args), body=body)
-    # TODO userId
+    sess = setup_session(args)
+    user_id = None
+    if args.user:
+        user_id = usernames_to_user_ids(sess, [args.user])[0]
+
+    body = bindings.v1GetGroupsRequest(offset=args.offset, limit=args.limit, userId=user_id)
+    resp = bindings.post_GetGroups(sess, body=body)
     if args.json:
         print(json.dumps(resp.to_json(), indent=2))
     else:
         if resp.groups is None:
             resp.groups = []
+        group_list = []
+        for g in resp.groups:
+            group = g.group.to_json()
+            group["numMembers"] = g.numMembers
+            group_list.append(group)
+
         render.render_objects(
-            v1GroupHeaders, [render.unmarshal(v1GroupHeaders, g.to_json()) for g in resp.groups]
+            v1GroupHeaders, [render.unmarshal(v1GroupHeaders, g) for g in group_list]
         )
 
 
 @authentication.required
+@require_feature_flag("rbacEnabled", rbac_flag_disabled_message)
 def describe_group(args: Namespace) -> None:
     session = setup_session(args)
     group_id = group_name_to_group_id(session, args.group_name)
@@ -74,6 +86,7 @@ def describe_group(args: Namespace) -> None:
 
 
 @authentication.required
+@require_feature_flag("rbacEnabled", rbac_flag_disabled_message)
 def add_user_to_group(args: Namespace) -> None:
     session = setup_session(args)
     usernames = args.usernames.split(",")
@@ -89,6 +102,7 @@ def add_user_to_group(args: Namespace) -> None:
 
 
 @authentication.required
+@require_feature_flag("rbacEnabled", rbac_flag_disabled_message)
 def remove_user_from_group(args: Namespace) -> None:
     session = setup_session(args)
     usernames = args.usernames.split(",")
@@ -104,6 +118,7 @@ def remove_user_from_group(args: Namespace) -> None:
 
 
 @authentication.required
+@require_feature_flag("rbacEnabled", rbac_flag_disabled_message)
 def change_group_name(args: Namespace) -> None:
     session = setup_session(args)
     group_id = group_name_to_group_id(session, args.old_group_name)
@@ -115,11 +130,19 @@ def change_group_name(args: Namespace) -> None:
 
 
 @authentication.required
+@require_feature_flag("rbacEnabled", rbac_flag_disabled_message)
 def delete_group(args: Namespace) -> None:
-    session = setup_session(args)
-    group_id = group_name_to_group_id(session, args.group_name)
-    bindings.delete_DeleteGroup(session, groupId=group_id)
-    print(f"user group with name {args.group_name} and ID {group_id} deleted")
+    if args.yes or render.yes_or_no(
+        "Deleting a group will result in an unrecoverable \n"
+        "deletion of the group along with all the membership  \n"
+        "information of the group. Do you still wish to proceed? \n"
+    ):
+        session = setup_session(args)
+        group_id = group_name_to_group_id(session, args.group_name)
+        bindings.delete_DeleteGroup(session, groupId=group_id)
+        print(f"user group with name {args.group_name} and ID {group_id} deleted")
+    else:
+        print("Skipping group deletion.")
 
 
 def usernames_to_user_ids(session: session.Session, usernames: List[str]) -> List[int]:
@@ -145,12 +168,12 @@ def usernames_to_user_ids(session: session.Session, usernames: List[str]) -> Lis
 
 
 def group_name_to_group_id(session: session.Session, group_name: str) -> int:
-    body = bindings.v1GroupSearchRequest(name=group_name)
+    body = bindings.v1GetGroupsRequest(name=group_name, limit=1, offset=0)
     resp = bindings.post_GetGroups(session, body=body)
     groups = resp.groups
-    if groups is None or len(groups) != 1 or groups[0].groupId is None:
+    if groups is None or len(groups) != 1 or groups[0].group.groupId is None:
         raise api.errors.BadRequestException(f"could not find user group name {group_name}")
-    return groups[0].groupId
+    return groups[0].group.groupId
 
 
 args_description = [
@@ -180,6 +203,7 @@ args_description = [
                 "deletes a user group",
                 [
                     Arg("group_name", default=None, help="name of user group to be deleted"),
+                    Arg("--yes", action="store_true", help="skip prompt asking for confirmation"),
                 ],
             ),
             Cmd(
@@ -188,6 +212,7 @@ args_description = [
                 "list user groups",
                 [
                     *default_pagination_args,
+                    Arg("--user", help="list groups that the username is in"),
                     Arg("--json", action="store_true", help="print as JSON"),
                 ],
                 is_default=True,
